move distirbution to separate file
